+++ lib/rc.dart
@@ -1,1 +1,3 @@
-export 'src/runtime_configuration.dart' show RuntimeConfiguration;
+export 'src/runtime_configuration.dart'
+    if (dart.library.io) 'src/runtime_configuration_io.dart';
+export 'src/runtime_context.dart';