+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+- Remove `package:rc/io.dart`
+- Add `RuntimeConfiguration.from` factory constructor
+- Add `all` to getter
+- Removed `load` method, now create `RuntimeConfiguration` will be resolved immediately
+
 ## 0.0.2
 
 Fix child replace throw error